pyproject.toml
[project]
name = "deep-agent-study"
version = "0.1.0"
description = "Deep Agent streaming debug with Azure OpenAI"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.4.10",
    "langchain-openai>=0.3.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]
